requirements.txt
# Core dependencies for AdaniGreenPredictor
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
feedparser>=6.0.10
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.17.0
scikit-learn>=1.3.0
torch>=2.0.0
pytorch-lightning>=2.1.0
pytorch-forecasting>=1.0.1
transformers>=4.35.0
sentencepiece>=0.1.99
python-dotenv>=1.0.0
accelerate>=0.24.0

